requirements.txt
transformers==4.49.0
torch>=2.1.0
Pillow>=10.0.0
PyMuPDF>=1.23.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
gradio>=4.20.0
einops>=0.7.0
timm>=0.9.0
